+++ api/models/Subscription.go
@@ -17,7 +17,7 @@ type Subscription struct {
 	Email     string    `gorm:"size:255;not null;" json:"email"`
 	Data      string    `gorm:"type:JSONB;not null;" json:"data"`
 	Author    User      `json:"author"`
-	AuthorID  uint32    `gorm:"not null" json:"author_id"`
+	AuthorID  uint64    `gorm:"not null" json:"author_id"`
 	CreatedAt time.Time `gorm:"default:CURRENT_TIMESTAMP" json:"created_at"`
 	UpdatedAt time.Time `gorm:"default:CURRENT_TIMESTAMP" json:"updated_at"`
 }
@@ -114,7 +114,7 @@ func (p *Subscription) UpdateASubscription(db *gorm.DB) (*Subscription, error) {
 }
 
 // DeleteASubscription - Удаление подписки
-func (p *Subscription) DeleteASubscription(db *gorm.DB, pid uint64, uid uint32) (int64, error) {
+func (p *Subscription) DeleteASubscription(db *gorm.DB, pid uint64, uid uint64) (int64, error) {
 	db = db.Debug().Model(&Subscription{}).Where("id = ? and author_id = ?", pid, uid).Take(&Subscription{}).Delete(&Subscription{})
 	if db.Error != nil {
 		if gorm.IsRecordNotFoundError(db.Error) {
